Move article related components into a subfolder
